+++ pixelwall.c
@@ -14,12 +14,7 @@ TERMS AND CONDITIONS FOR COPYING, DISTRIBUTION AND MODIFICATION
 0. You just DO WHAT THE FUCK YOU WANT TO.
 */
 
-#include "raylib.h"
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>  
-#include <unistd.h>
-
+#include "design.h"
 
 // Configuration
 #define DEFAULT_ROWS 16
@@ -50,40 +45,6 @@ Color BACKGROUND_COLOR = DEFAULT_BACKGROUND_COLOR;
 Color WORM_COLOR = DEFAULT_WORM_COLOR;
 Color BORDER_COLOR = DEFAULT_BORDER_COLOR;
 
-// Add Pixel struct definition
-typedef struct {
-    Color color;
-    bool isWorm;
-    bool isFruit;
-} Pixel;
-
-// Game objects
-typedef struct {
-    Vector2 position;
-    Color color;
-} Segment;
-
-typedef struct {
-    Vector2 position;
-    Color color;
-} Fruit;
-
-typedef struct {
-    Segment *body;      // Pointer to dynamically allocated array
-    int length;         // Current length of the worm
-    int max_length;     // Maximum allocated length
-    Vector2 direction;  // Current movement direction
-} Worm;
-
-// Main game state structure
-typedef struct {
-    Pixel **grid;        // 2D array representing the game grid
-    Worm worm;           // The snake/worm object
-    Fruit fruit;         // The fruit object
-    float moveTimer;     // Timer for movement updates
-    Vector2 currentDir;  // Current movement direction
-    bool gameOver;       // Game over flag
-} GameState;
 
 // Function prototypes
 void InitializeGrid(GameState *state);
@@ -98,6 +59,18 @@ void CleanupWorm(GameState *state);
 
 int cellWidth;
 int cellHeight;
+
+// Initialize the game grid with default values
+void InitializeGrid(GameState *state) {
+    // Loop through all rows and columns
+    for (int row = 0; row < ROWS; row++) {
+        for (int col = 0; col < COLS; col++) {
+            state->grid[row][col].color = BACKGROUND_COLOR;  // Set background color
+            state->grid[row][col].isWorm = false;            // No worm initially
+            state->grid[row][col].isFruit = false;           // No fruit initially
+        }
+    }
+}
 
 // Add helper function for random colors
 Color GetRandomColor() {
@@ -120,119 +93,6 @@ Vector2 GetRandomDirection() {
     return dirs[GetRandomValue(0, 3)];
 }
 
-bool IsPositionValid(const GameState *state, Vector2 pos) {
-    return pos.x >= 0 && pos.x < COLS && 
-           pos.y >= 0 && pos.y < ROWS &&
-           !state->grid[(int)pos.y][(int)pos.x].isWorm;
-}
-
-Vector2 CalculateNewHead(const GameState *state) {
-    Vector2 head = state->worm.body[0].position;
-    return (Vector2){head.x + state->currentDir.x, 
-                    head.y + state->currentDir.y};
-}
-
-// Update worm position and handle growth
-void UpdateWormPosition(GameState *state, Vector2 newHead, bool growing) {
-    // If growing, increment length before updating positions
-    if (growing) {
-        state->worm.length++;
-    } else {
-        // Clear the old tail position only if not growing
-        Vector2 oldTail = state->worm.body[state->worm.length - 1].position;
-        state->grid[(int)oldTail.y][(int)oldTail.x].color = BACKGROUND_COLOR;
-        state->grid[(int)oldTail.y][(int)oldTail.x].isWorm = false;
-    }
-    
-    // Move worm segments forward
-    for (int i = state->worm.length - 1; i > 0; i--) {
-        state->worm.body[i] = state->worm.body[i - 1];
-    }
-    state->worm.body[0].position = newHead;
-    
-    // Update grid with new positions
-    for (int i = 0; i < state->worm.length; i++) {
-        Vector2 pos = state->worm.body[i].position;
-        state->grid[(int)pos.y][(int)pos.x].color = WORM_COLOR;
-        state->grid[(int)pos.y][(int)pos.x].isWorm = true;
-    }
-}
-
-// Handle fruit collision and create new fruit
-void HandleFruitCollision(GameState *state) {
-    // Increase worm length if possible
-    if (state->worm.length < MAX_WORM_LENGTH) {
-        // Add new segment at the end
-        state->worm.body[state->worm.length] = state->worm.body[state->worm.length - 1];
-        state->worm.length++;
-    }
-    
-    // Remove old fruit from grid
-    Vector2 oldFruitPos = state->fruit.position;
-    state->grid[(int)oldFruitPos.y][(int)oldFruitPos.x].isFruit = false;
-    
-    // Create new fruit at random position
-    InitializeFruit(state);
-}
-
-// Helper function to check if a move is safe
-bool IsSafeMove(const GameState *state, Vector2 pos) {
-    return pos.x >= 0 && pos.x < COLS && 
-           pos.y >= 0 && pos.y < ROWS && 
-           !state->grid[(int)pos.y][(int)pos.x].isWorm;
-}
-
-// Simplified UpdateAI function
-void UpdateAI(GameState *state) {
-    Vector2 head = state->worm.body[0].position;
-    Vector2 fruitPos = state->fruit.position;
-    
-    // First try to move towards fruit
-    Vector2 desiredDir = {0, 0};
-    
-    // Try horizontal movement first
-    if (fruitPos.x != head.x) {
-        desiredDir.x = (fruitPos.x > head.x) ? 1 : -1;
-        desiredDir.y = 0;
-    } 
-    // Then try vertical movement
-    else if (fruitPos.y != head.y) {
-        desiredDir.x = 0;
-        desiredDir.y = (fruitPos.y > head.y) ? 1 : -1;
-    }
-    
-    // Check if desired move is safe
-    Vector2 nextPos = {
-        head.x + desiredDir.x,
-        head.y + desiredDir.y
-    };
-    
-    if (IsSafeMove(state, nextPos)) {
-        state->currentDir = desiredDir;
-        return;
-    }
-    
-    // If desired move is not safe, try other directions
-    Vector2 directions[] = {
-        {1, 0},   // Right
-        {-1, 0},  // Left
-        {0, 1},   // Down
-        {0, -1}   // Up
-    };
-    
-    for (int i = 0; i < 4; i++) {
-        Vector2 dir = directions[i];
-        Vector2 newPos = {head.x + dir.x, head.y + dir.y};
-        
-        if (IsSafeMove(state, newPos)) {
-            state->currentDir = dir;
-            return;
-        }
-    }
-    
-    // If no safe moves found, keep current direction (will trigger game over)
-}
-
 // Main game loop
 int main(int argc, char *argv[]) {  // Change main signature to standard form
     ParseCommandLine(argc, argv);   // Use standard argc/argv instead of __argc/__argv
@@ -246,49 +106,16 @@ int main(int argc, char *argv[]) {  // Change main signature to standard form
 
     // Initialize game state
     GameState state;
+    
     InitializeGridMemory(&state);
     InitializeGrid(&state);
-    InitializeWorm(&state);
-    InitializeFruit(&state);
-    
-    state.moveTimer = 0.0f;  // Initialize movement timer
-    state.currentDir = GetRandomDirection();  // Set initial direction
-    state.gameOver = false;  // Game is running
+
+    DesignInit(&state);
 
     // Main game loop
     while (!WindowShouldClose()) {
         // Update game state
-        state.moveTimer += GetFrameTime();
-        if (state.moveTimer >= MOVE_INTERVAL) {
-            if (state.gameOver) {
-                // Reset game state if game over
-                InitializeGrid(&state);
-                InitializeWorm(&state);
-                InitializeFruit(&state);
-                state.currentDir = GetRandomDirection();
-                state.gameOver = false;
-            }
-            else {
-                // Update AI and check for collisions
-                UpdateAI(&state);
-                
-                Vector2 newHead = CalculateNewHead(&state);
-                if (CheckGameOver(&state, newHead)) {
-                    state.gameOver = true;
-                }
-                else {
-                    // Check if the worm is about to eat fruit
-                    bool willGrow = state.grid[(int)newHead.y][(int)newHead.x].isFruit;
-                    
-                    UpdateWormPosition(&state, newHead, willGrow);
-                    
-                    if (willGrow) {
-                        HandleFruitCollision(&state);
-                    }
-                }
-            }
-            state.moveTimer = 0.0f;  // Reset movement timer
-        }
+        DesignUpdateFrame(&state);
         
         // Draw game state
         BeginDrawing();
@@ -302,60 +129,6 @@ int main(int argc, char *argv[]) {  // Change main signature to standard form
     CleanupGridMemory(&state);
     CloseWindow();
     return 0;
-}
-
-// Initialize the game grid with default values
-void InitializeGrid(GameState *state) {
-    // Loop through all rows and columns
-    for (int row = 0; row < ROWS; row++) {
-        for (int col = 0; col < COLS; col++) {
-            state->grid[row][col].color = BACKGROUND_COLOR;  // Set background color
-            state->grid[row][col].isWorm = false;            // No worm initially
-            state->grid[row][col].isFruit = false;           // No fruit initially
-        }
-    }
-}
-
-// Initialize the worm at the center of the grid
-void InitializeWorm(GameState *state) {
-    state->worm.max_length = MAX_WORM_LENGTH;
-    state->worm.body = (Segment *)malloc(state->worm.max_length * sizeof(Segment));
-    if (!state->worm.body) {
-        fprintf(stderr, "Failed to allocate memory for worm body\n");
-        exit(EXIT_FAILURE);
-    }
-    state->worm.length = WORM_LENGTH;
-    // Start worm in the middle of the grid
-    int startRow = ROWS / 2;
-    int startCol = COLS / 2;
-    
-    // Initialize worm segments from head to tail
-    for (int i = 0; i < WORM_LENGTH; i++) {
-        state->worm.body[i].position = (Vector2){startCol - i, startRow};
-        state->worm.body[i].color = WORM_COLOR;
-        
-        // Mark positions in grid
-        int col = startCol - i;
-        int row = startRow;
-        state->grid[row][col].color = WORM_COLOR;
-        state->grid[row][col].isWorm = true;
-    }
-}
-
-void InitializeFruit(GameState *state) {
-    // Place fruit in random position (not on worm)
-    int row, col;
-    do {
-        row = GetRandomValue(0, ROWS - 1);
-        col = GetRandomValue(0, COLS - 1);
-    } while (state->grid[row][col].isWorm);
-    
-    state->fruit.position = (Vector2){col, row};
-    state->fruit.color = GetRandomColor();
-    
-    // Mark position in grid
-    state->grid[row][col].color = state->fruit.color;
-    state->grid[row][col].isFruit = true;
 }
 
 void DrawPixelGrid(const GameState *state) {
@@ -412,11 +185,6 @@ void CleanupGridMemory(GameState *state) {
         }
         free(state->grid);
     }
-}
-
-bool CheckGameOver(const GameState *state, Vector2 newHead) {
-    // Check if new head position is invalid
-    return !IsPositionValid(state, newHead);
 }
 
 void PrintHelp() {
@@ -516,33 +284,3 @@ void ParseCommandLine(int argc, char *argv[]) {
     }
 }
 
-void CleanupWorm(GameState *state) {
-    if (state->worm.body) {
-        free(state->worm.body);
-        state->worm.body = NULL;
-    }
-    state->worm.length = 0;
-    state->worm.max_length = 0;
-}
-
-void GrowWorm(Worm *worm) {
-    if (worm->length >= worm->max_length) {
-        // Handle maximum length reached
-        return;
-    }
-    // Add new segment
-    worm->length++;
-}
-
-void ResizeWorm(Worm *worm, int new_max_length) {
-    Segment *new_body = (Segment *)realloc(worm->body, new_max_length * sizeof(Segment));
-    if (!new_body) {
-        fprintf(stderr, "Failed to resize worm body\n");
-        return;
-    }
-    worm->body = new_body;
-    worm->max_length = new_max_length;
-    if (worm->length > worm->max_length) {
-        worm->length = worm->max_length;
-    }
-}
